pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "multibench"
version = "0.1.0"
description = "Self-contained evaluation harness for personalization benchmarks over vLLM / OpenAI-compatible endpoints."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Mixed (per-benchmark upstream licenses retained)" }
dependencies = [
    "openai>=1.30.0",
    "tqdm>=4.66",
    "pyyaml>=6.0",
    "tiktoken>=0.5",
    "diskcache>=5.6",
    "beautifulsoup4>=4.12",
    "evaluate>=0.4",
    "sacrebleu>=2.4",
    "rouge-score>=0.1.2",
    "nltk>=3.9",
    "scipy>=1.11",
    "pandas>=2.0",
    "numpy>=1.24",
    "scikit-learn>=1.3",
]

[project.optional-dependencies]
bedrock = ["boto3>=1.34"]
gemini  = ["google-genai>=1.0"]
replicate = ["replicate>=0.30"]

[project.scripts]
multibench = "multibench.cli:main"

[tool.setuptools.packages.find]
include = ["multibench*"]
